[mod] 몬스터도 무기 데미지 계산으로 0 이하일 때 죽도록 수정. 충돌 후 일정시간동안 무적 적용 해야할 것 무기 스테이터스와 플레이어 스테이터스 수식 적용

+++ SurviveGame/Assets/Scripts/Controller/Player/PlayerController.cs
@@ -100,7 +100,7 @@ public class PlayerController : MonoBehaviour , ITargetAble
         return playerBoxInfo;
     }
 
-    public void OnDamege(int _damageAmount)
+    public void OnDamege(float _damageAmount)
     {
         curHP -= _damageAmount;
         float fillAmount = curHP / character.statusArr[(int)StatusEffectType.P_MaxHP].status;
